Make Host Creation Non Blocking
I recently introduced --noautoconsole as it makes sense.  This however made creation non-blocking
and virt-install was subsequently unable to reboot the instance once it shut down.  This extends
the API to allow a host to be started.  This also renames host 'find' to be more intuitively 'get'
and also checks if the host is shut off before attempting to destroy it.